Room Coloring a Matrix
The program will request the width and length of the matrix. Then it will request a matrix will 0 and 1's and one of the element's in the matrix will be a colour. (Ex. B for Blue, G for Green, etc.). The program will color all the zero's as the color put into the matrix.